# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the CraftedBy API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a list of products",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "description": "Creates a new product",
                "operationId": "addProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product added"
                    },
                    "default": {
                        "description": "Product not added, oops"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Find product by ID",
                "description": "Returns a single product",
                "operationId": "7ab21ddb713501f3d4c2914892c02a21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of product to return",
                        "required": true,
                        "schema": {
                            "description": "A UUID string with 36 characters",
                            "type": "string",
                            "format": "uuid",
                            "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/api/products/shop/{idshop}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all the products from 1 shop",
                "operationId": "e37e033f62ba1d7262b82523a5076f1f",
                "parameters": [
                    {
                        "name": "idshop",
                        "in": "path",
                        "description": "ID of the shop",
                        "required": true,
                        "schema": {
                            "description": "A UUID string with 36 characters",
                            "type": "string",
                            "format": "uuid",
                            "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/api/products/{category}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all the products from 1 category",
                "operationId": "b93b8da7f89613b3a7912162963c40a2",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/products/{price}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all the products of a price",
                "operationId": "d79c64a45eea80afa8dd3925c2a6e161",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/products/{color}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all the products of a color",
                "operationId": "4abe34d8bee6e8ffdfc16ba63abdd022",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/products/{material}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all the products of a material",
                "operationId": "581e7a3b8f89e118d95297369ebbee1f",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/products/{size}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all the products of a size",
                "operationId": "ccd8ca44aa678a43d154fc007b66bf9f",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/products/{name}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all the products of a name",
                "operationId": "fb9be688953226e158fe4f76326be407",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "quantity": {
                        "type": "number"
                    },
                    "material": {
                        "type": "string"
                    },
                    "color": {
                        "type": "string"
                    },
                    "size": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string"
                    },
                    "img_url": {
                        "type": "string"
                    },
                    "shop_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    }
}